Add JS support to node package
