+++ host-go/node/node.go
@@ -10,23 +10,14 @@ import (
 
 	badgerds "github.com/dgraph-io/badger/v4"
 	"github.com/sourcenetwork/corekv"
-	sourceP2P "github.com/sourcenetwork/go-p2p"
 
 	"github.com/sourcenetwork/corekv/badger"
 	"github.com/sourcenetwork/immutable"
-	"github.com/sourcenetwork/lens/host-go/p2p"
 	"github.com/sourcenetwork/lens/host-go/runtimes"
 	"github.com/sourcenetwork/lens/host-go/store"
 )
 
 type closer = func() error
-
-type Node struct {
-	onClose []closer
-	Options Options
-	Store   store.TxnStore
-	P2P     immutable.Option[*p2p.P2P]
-}
 
 func New(ctx context.Context, opts ...Option) (*Node, error) {
 	var o Options
@@ -77,39 +68,9 @@ func New(ctx context.Context, opts ...Option) (*Node, error) {
 		o.IndexstoreNamespace = immutable.Some("i")
 	}
 
-	var p2pSys immutable.Option[*p2p.P2P]
-	if !o.DisableP2P {
-		var host p2p.Host
-		if o.P2P.HasValue() {
-			host = o.P2P.Value()
-		} else {
-			p2pOptions := append(
-				o.P2POptions,
-				sourceP2P.WithBlockstoreNamespace(o.BlockstoreNamespace.Value()),
-				sourceP2P.WithRootstore(o.Rootstore.Value()),
-			)
-
-			if o.BlockstoreChunkSize.HasValue() {
-				p2pOptions = append(p2pOptions, sourceP2P.WithBlockstoreChunkSize(o.BlockstoreChunkSize.Value()))
-			}
-
-			var err error
-			host, err = sourceP2P.NewPeer(
-				ctx,
-				p2pOptions...,
-			)
-			if err != nil {
-				return nil, err
-			}
-		}
-
-		p2pSys = immutable.Some(p2p.New(host, o.Rootstore.Value(), o.IndexstoreNamespace.Value()))
-	}
-
-	node := &Node{
-		onClose: onClose,
-		Options: o,
-		Store: store.New(
+	node, err := createNode(
+		ctx,
+		store.New(
 			o.TxnSource.Value(),
 			o.PoolSize.Value(),
 			o.Runtime.Value(),
@@ -117,12 +78,16 @@ func New(ctx context.Context, opts ...Option) (*Node, error) {
 			o.BlockstoreChunkSize,
 			o.IndexstoreNamespace.Value(),
 		),
-		P2P: p2pSys,
+		o,
+		onClose,
+	)
+	if err != nil {
+		return nil, err
 	}
 
 	// Reload on create, if the store is persisted, this will read any Lenses already in the store and
 	// add them to the repository.
-	err := node.Store.Reload(ctx)
+	err = node.Store.Reload(ctx)
 	if err != nil {
 		return nil, err
 	}
