+++ host-go/node/option.go
@@ -6,27 +6,10 @@ package node
 
 import (
 	"github.com/sourcenetwork/corekv"
-	sourceP2P "github.com/sourcenetwork/go-p2p"
 	"github.com/sourcenetwork/immutable"
 	"github.com/sourcenetwork/lens/host-go/engine/module"
-	"github.com/sourcenetwork/lens/host-go/p2p"
 	"github.com/sourcenetwork/lens/host-go/store"
 )
-
-type Options struct {
-	Path                immutable.Option[string]
-	Rootstore           immutable.Option[corekv.TxnReaderWriter]
-	TxnSource           immutable.Option[store.TxnSource]
-	PoolSize            immutable.Option[int]
-	Runtime             immutable.Option[module.Runtime]
-	BlockstoreNamespace immutable.Option[string]
-	BlockstoreChunkSize immutable.Option[int]
-	IndexstoreNamespace immutable.Option[string]
-	P2P                 immutable.Option[p2p.Host]
-	DisableP2P          bool
-
-	P2POptions []sourceP2P.NodeOpt
-}
 
 // Option is a funtion that sets a config value on the db.
 type Option func(*Options)
@@ -66,24 +49,6 @@ func WithBlockstoreNamespace(blockstoreNamespace string) Option {
 func WithIndextoreNamespace(indexstoreNamespace string) Option {
 	return func(opts *Options) {
 		opts.IndexstoreNamespace = immutable.Some(indexstoreNamespace)
-	}
-}
-
-func WithP2P(host p2p.Host) Option {
-	return func(opts *Options) {
-		opts.P2P = immutable.Some(host)
-	}
-}
-
-func WithP2PDisabled(disableP2P bool) Option {
-	return func(opts *Options) {
-		opts.DisableP2P = disableP2P
-	}
-}
-
-func WithP2Poptions(opts ...sourceP2P.NodeOpt) Option {
-	return func(opt *Options) {
-		opt.P2POptions = opts
 	}
 }
